Refact: Rename ...commandexecution package to just ...command
