+++ examples/google/src/main/java/io/github/arthurc/example/autotest/google/GooglePage.java
@@ -4,7 +4,7 @@
 package io.github.arthurc.example.autotest.google;
 
 import io.github.arthurc.autotest.AutotestScoped;
-import io.github.arthurc.autotest.commandexecution.Command;
+import io.github.arthurc.autotest.command.Command;
 import io.github.arthurc.autotest.web.Browser;
 import org.springframework.stereotype.Component;
